+++ CHANGELOG.md
@@ -19,6 +19,7 @@
     * `a ? a : b` is minified to `a || b`
     * `a == void 0` is minified to `a == null`
     * `a && (b && c)` is minified to `a && b && c` (same for `||`)
+    * `(function foo() {})` is minified to `(function() {})`
     * `typeof a === "string"` is minified to `typeof a == "string"`
     * `if (a) if (b) return c` is minified to `if (a && b) return c`
     * `while (a) if (!b) break;` is minified to `for (; a && b; ) ;`